[fix #75] OrderController 테스트 메서드 추가
