+++ docker-compose.yml
@@ -12,6 +12,11 @@ services:
       retries: 5
       start_period: 30s
       timeout: 10s
+    command: [
+      "sh",
+      "-c",
+      "redis-server --notify-keyspace-events Ex"
+    ]
 
   user-db:
     container_name: user-db
