+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/infrastructure/redis/RedissonConfig.java
@@ -1,21 +1,25 @@
 package com.ticketPing.queue_manage.infrastructure.redis;
 
-import static com.ticketPing.queue_manage.infrastructure.utils.YmlLoader.redisHost;
-import static com.ticketPing.queue_manage.infrastructure.utils.YmlLoader.redisPort;
-
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
 import org.redisson.config.Config;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class RedissonConfig {
 
+    @Value("${redis.host}")
+    private String redisHost;
+
+    @Value("${redis.port}")
+    private String redisPort;
+
     @Bean
     public RedissonClient redissonClient() {
         Config config = new Config();
-        config.useSingleServer().setAddress(String.format("redis://%s:%s", redisHost(), redisPort()));
+        config.useSingleServer().setAddress(String.format("redis://%s:%s", redisHost, redisPort));
         return Redisson.create(config);
     }
